requirements.txt
PyTorch
torchvision
timm
yacs
opencv-python
scikit-learn
faiss-cpu
numpy
Pillow
